pyproject.toml
[tool.poetry]
name = "shape_of_sort"
version = "0.1.0"
description = ""
authors = ["Paul Albee <paula@allenai.org>"]

[tool.poetry.dependencies]
python = "~3.10"
ipython = "^8.4.0"
numpy = "^1.23.2"
matplotlib = "^3.5.3"
scikit-image = "^0.19.3"
Pillow = "^9.2.0"
tabulate = "^0.8.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
